_files/requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
pillow==10.2.0
openpyxl==3.1.2
pystray==0.19.5
